+++ CHANGELOG.md
@@ -26,7 +26,9 @@ and this project adheres to
   [#2365](https://github.com/OpenFn/lightning/issues/2365)
 - User email change: Send notification of change to the old email address and
   confirmation to the new email address [#2365](https://github.com/OpenFn/lightning/issues/2365)
-  
+- Export edge condition expressions as multiline in project spec
+  [#2521](https://github.com/OpenFn/lightning/issues/2521)
+
 ## [v2.9.6] - 2024-09-23
 
 ### Added